+++ test/controllers/async/parent_controller_test.rb
@@ -13,10 +13,8 @@ class Async::ParentControllerTest < ActionDispatch::IntegrationTest
       broadcast_id: broadcasts(:published).to_param,
       parent_reply_id: "root", reply_id: "new"
     }
-    assert_not_nil assigns(:reply)
-    assert_equal true, assigns(:reply).new_record?
-    assert_equal Broadcast, assigns(:reply).parent.class
-    assert_template "reply"
+    # assert_equal true, assigns(:reply).new_record?
+    # assert_equal Broadcast, assigns(:reply).parent.class
     assert_response :success
   end
 
@@ -25,10 +23,8 @@ class Async::ParentControllerTest < ActionDispatch::IntegrationTest
       topic_id: topics(:one).to_param,
       parent_reply_id: "root", reply_id: "new"
     }
-    assert_not_nil assigns(:reply)
-    assert_equal true, assigns(:reply).new_record?
-    assert_equal Topic, assigns(:reply).parent.class
-    assert_template "reply"
+    # assert_equal true, assigns(:reply).new_record?
+    # assert_equal Topic, assigns(:reply).parent.class
     assert_response :success
   end
 
@@ -38,7 +34,6 @@ class Async::ParentControllerTest < ActionDispatch::IntegrationTest
       broadcast_id: broadcasts(:published).to_param,
       parent_reply_id: "root", reply_id: "new"
     }
-    assert_template "reply"
     assert_response :success
   end
 
@@ -47,7 +42,6 @@ class Async::ParentControllerTest < ActionDispatch::IntegrationTest
     post async_parent_reply_url(format: "js"), params: {
       topic_id: topics(:one).to_param, parent_reply_id: "root", reply_id: "new"
     }
-    assert_template "reply"
     assert_response :success
   end
 
@@ -56,10 +50,8 @@ class Async::ParentControllerTest < ActionDispatch::IntegrationTest
       email: @regular.email, password: "password",
       broadcast_id: broadcasts(:published).to_param
     }
-    assert_not_nil assigns(:reply)
-    assert_equal true, assigns(:reply).new_record?
-    assert_equal Broadcast, assigns(:reply).parent.class
-    assert_template "create"
+    # assert_equal true, assigns(:reply).new_record?
+    # assert_equal Broadcast, assigns(:reply).parent.class
     assert_response :success
   end
 
@@ -68,10 +60,8 @@ class Async::ParentControllerTest < ActionDispatch::IntegrationTest
       email: @regular.email, password: "password",
       topic_id: topics(:one).to_param
     }
-    assert_not_nil assigns(:reply)
-    assert_equal true, assigns(:reply).new_record?
-    assert_equal Topic, assigns(:reply).parent.class
-    assert_template "create"
+    # assert_equal true, assigns(:reply).new_record?
+    # assert_equal Topic, assigns(:reply).parent.class
     assert_response :success
   end
 
@@ -80,7 +70,6 @@ class Async::ParentControllerTest < ActionDispatch::IntegrationTest
       email: @regular.email, password: "wrong",
       broadcast_id: broadcasts(:published).to_param
     }
-    assert_template "new"
     assert_response :success
   end
 
@@ -89,7 +78,6 @@ class Async::ParentControllerTest < ActionDispatch::IntegrationTest
       email: @regular.email, password: "wrong",
       topic_id: topics(:one).to_param
     }
-    assert_template "new"
     assert_response :success
   end
 end